face_project/server/requirements.txt
tornado==5.1.1
pillow==5.3.0
opencv-python==3.4.10.35
numpy==1.19.5
WTForms==2.2.1
Werkzeug==0.14.1